Bug 1244049 - Part 1: Define scoped enum for CSSPseudoElement type. r=dbaron
UltraBlame original commit: 4a020e2e4acec77f56e1ccea319765a7e18c74c9

+++ dom/animation/PseudoElementHashEntry.h
@@ -47,7 +47,9 @@ public:
     if (!aKey)
       return 0;
 
-    return mozilla::HashGeneric(aKey->mElement, aKey->mPseudoType);
+    
+    return mozilla::HashGeneric(aKey->mElement,
+                                static_cast<uint8_t>(aKey->mPseudoType));
   }
   enum { ALLOW_MEMMOVE = true };
